Creation of pool of sounds done

+++ src/GA/Mutation.java
@@ -4,12 +4,12 @@ import java.util.Random;
 
 public class Mutation {
 
-    public static Genotype getRandomGenes(int maxKey, int maxVelocity, int maxTick){
+    public static Genotype getRandomGenes(int maxKey, int maxVelocity, long maxTick){
         Random random = new Random();
         Genotype genotype = new Genotype();
         genotype.setKey(random.nextInt(maxKey));
         genotype.setVelocity(random.nextInt(maxVelocity)+1);
-        genotype.setTick(random.nextInt(maxTick));
+        genotype.setTick(random.nextInt((int)maxTick));
         genotype.setNote(random.nextBoolean());
         return genotype;
     }
