+++ src/GA/Populate.java
@@ -3,7 +3,7 @@ package GA;
 import java.util.ArrayList;
 
 public class Populate {
-    public static ArrayList<Individual> initPool(int poolSize, int popSize, int maxKey, int maxVelocity, int maxTick){
+    public static ArrayList<Individual> initPool(int poolSize, int popSize, int maxKey, int maxVelocity, long maxTick){
         ArrayList<Individual> soundPopulation = new ArrayList<>();
         for(int i = 0; i < poolSize; i++){
             ArrayList<Genotype> genes = new ArrayList<>();
